% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-Date-E-addPeriodToDate.R
\name{addPeriodToDate}
\alias{addPeriodToDate}
\title{function to add or subtract periods to dates}
\usage{
addPeriodToDate(initialDate, period, numPeriods)
}
\arguments{
\item{initialDate}{valid date in the YYYY-MM-DD format}

\item{period}{a character string, one of "days", "weeks", "months" or "years"}

\item{numPeriods}{an integer (can be negative) specifying the number of periods to add (or subtract)}
}
\value{
a date; note: this will account for leap years, if Feb 29 is passed, it will default to Feb 28th for non-leap years
}
\description{
function to add or subtract periods to dates
}
\examples{
none

}

