% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-computeDrawdownMatrix.R
\name{computeDrawdownMatrix}
\alias{computeDrawdownMatrix}
\title{compute drawdown for time series matrix}
\usage{
computeDrawdownMatrix(timeSeriesMatrix, continuous = FALSE)
}
\arguments{
\item{continuous}{logical; if true, will compute index levels for continuous returns, if FALSE, will compute index levels for linear returns; defaults to FALSE}

\item{timeSeriesVector}{xts time series matrix}
}
\value{
a matrix of drawdown levels
}
\description{
compute drawdown for time series matrix
}
\examples{
none

}

