% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-Estimation-E-COVtoCOR.R
\name{COVtoCOR}
\alias{COVtoCOR}
\title{convert covariance matrix to correlation matrix}
\usage{
COVtoCOR(covarianceMatrix)
}
\arguments{
\item{covarianceMatrix}{}
}
\value{
correlation matrix
}
\description{
convert covariance matrix to correlation matrix
}
\examples{
none

}

