% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-Optimization-E-createOptimalStatisticTimeSeries.R
\name{createOptimalStatisticTimeSeries}
\alias{createOptimalStatisticTimeSeries}
\title{extract optimization statistics from the optimization object}
\usage{
createOptimalStatisticTimeSeries(optimizationResultsObject, statisticType,
  datePeriodType = "train", annualize = FALSE, indexType = "timeSeries",
  indexTimeSeriesFreq = "daily")
}
\arguments{
\item{optimizationResultsObject}{a list containing rolling statistics for an optimization}

\item{statisticType}{character string, one of "opHoldings", "opWeights", "opER", "opSD", "opSharpe", "opIndex" for Optimal result; one of "mvHoldings", "mvWeights", "mvER", "mvSD", "mvSharpe", "mvIndex" for Minimum Variance result; or for a list of Efficient Fronter numbers, "ER" for expected return, "SD" for standard deviation, and "EF" for a list of matrices where the first column is standard deviation, and second column is expected return}

\item{datePeriodType}{character string, one of "train", "test", currently assumes one month ahead, defaults to "train"}

\item{annualize}{logical that will annualize Expected Return, Standard Deviation or Sharpe ratio, if it's not applicable, it will be ignored; defaults to FALSE}

\item{indexType}{character string, one of "timeSeries" or "raw"; where "timeSeries" generates a daily xts object of the statistic, and "raw" appends the rolling time period to the statistic}

\item{indexTimeSeriesFreq}{character string, one of "daily", "weekly", "monthly", or "yearly"; only applicable if indexType = "timeSeries", defaults to "daily"}
}
\value{
an xts matrix of results OR a list of matrices depending on the selection
}
\description{
extract optimization statistics from the optimization object
}
\examples{
FUNCTION STILL UNDER DEVELOPMENT

}

