% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-Optimization-E-createRealizedPortfolioTimeSeries.R
\name{createRealizedPortfolioTimeSeries}
\alias{createRealizedPortfolioTimeSeries}
\title{extract optimization statistics from the optimization object}
\usage{
createRealizedPortfolioTimeSeries(linearReturnMatrix, weightMatrix,
  includeConstituents = FALSE, benchmark = NA,
  portfolioName = "Portfolio", benchmarkName = "Benchmark",
  rebalance = TRUE)
}
\arguments{
\item{linearReturnMatrix}{a linear return xts matrix}

\item{weightMatrix}{a weight xts object, with same number of columns as the linearReturnMatrix}

\item{includeConstituents}{logical, if TRUE, the output will append the unweighted constituents of the portfolio; defaults to FALSE}

\item{benchmark}{an xts vector of linear returns; if not NA, then will append this to the resulting xts object; defaults to NA (not inlcuded)}

\item{portfolioName}{a character string, defaults to "Portfolio"}

\item{benchmarkName}{a character string, used to rename a benchmark that's included; defaults to "Benchmark"}
}
\value{
an xts matrix of results OR a list of matrices depending on the selection
}
\description{
extract optimization statistics from the optimization object
}
\examples{
FUNCTION STILL UNDER DEVELOPMENT

}

