% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-Date-E-createRollingTimeIndices.R
\name{createRollingTimeIndices}
\alias{createRollingTimeIndices}
\title{create index of time-stamps for vectorized functions}
\usage{
createRollingTimeIndices(initialTimeVector, estimationLength = 18,
  timeUnit = "months", investmentHorizonLength = 1,
  windowType = "rolling")
}
\arguments{
\item{initialTimeVector}{vector of dates YYYY-MM-DD}

\item{estimationLength}{a positive integer indicating the size of the estimation period}

\item{timeUnit}{a character string, one of "weeks", "days","weeks","months","years", indicating the unit size for estimationLength and investmentHorizon}

\item{investmentHorizonLength}{a positive integer indicating the size of the investment horizon (out of sample period)}
}
\value{
a character matrix of xts-subsettable date ranges, one column for the in-sample periods, and another for the out-of-sample periods
}
\description{
create index of time-stamps for vectorized functions
}
\examples{
FUNCTION STILL UNDER DEVELOPMENT

}

