% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-Date-E-createRollingTimeIndicesNEWTEST.R
\name{createRollingTimeIndicesNEWTEST}
\alias{createRollingTimeIndicesNEWTEST}
\title{create index of time-stamps for vectorized functions}
\usage{
createRollingTimeIndicesNEWTEST(initialTimeVector, numPeriods = (52 * 5))
}
\arguments{
\item{initialTimeVector}{vector of dates YYYY-MM-DD}

\item{numPeriods}{default to 52*5 weeks}
}
\value{
vector of YYYY-MM dates
}
\description{
create index of time-stamps for vectorized functions
}
\examples{
FUNCTION STILL UNDER DEVELOPMENT

}

