% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-dominantEigenvectors.R
\name{dominantEigenvectors}
\alias{dominantEigenvectors}
\title{Identify dominant eigenvalues}
\usage{
dominantEigenvectors(eigenResult, fitThreshold = 0.3)
}
\arguments{
\item{eigenResult}{which is a list of eigenvalues and eigenvectors}

\item{fitThreshold}{for Marcenko Pasteur distribution which decides if RMT will be used or a simple "minimim distance" method (Scree-plot approach)}
}
\value{
a list object with eigenvalues and eigenvectors
}
\description{
Identify dominant eigenvalues
}
\examples{
none

}

