% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-eigenDimensionReduction.R
\name{eigenDimensionReduction}
\alias{eigenDimensionReduction}
\title{Wrapperfor base::eigen, used in case this function switches to Rcpp eigen}
\usage{
eigenDimensionReduction(covarianceMatrix)
}
\arguments{
\item{covarianceMatrix}{which is a numeric square matrix}
}
\value{
a list object with eigenvalues and eigenvectors
}
\description{
Wrapperfor base::eigen, used in case this function switches to Rcpp eigen
}
\examples{
none

}

