% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-Estimation-E-estimateCovarianceMatrix.R
\name{estimateCovarianceMatrix}
\alias{estimateCovarianceMatrix}
\title{detailed function to estimate robust covariance matrix based on some assumptions}
\usage{
estimateCovarianceMatrix(timeSeriesMatrix)
}
\arguments{
\item{timeSeriesMatrix}{}
}
\value{
robust covariance matrix
}
\description{
detailed function to estimate robust covariance matrix based on some assumptions
}
\examples{
none

}

