% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-evaluateNormalDistributions.R
\name{evaluateNormalDistributions}
\alias{evaluateNormalDistributions}
\title{eval normal distributions}
\usage{
evaluateNormalDistributions(timeSeriesMatrix, listOfParameters)
}
\arguments{
\item{timeSeriesMatrix}{description TBD}

\item{listOfParameters}{description TBD}
}
\value{
evaluated normal distribution fits
}
\description{
eval normal distributions
}
\examples{
none

}

