% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-evaluateSkewStudentTDistributions.R
\name{evaluateSkewStudentTDistributions}
\alias{evaluateSkewStudentTDistributions}
\title{eval skew t distributions}
\usage{
evaluateSkewStudentTDistributions(timeSeriesMatrix, listOfParameters)
}
\arguments{
\item{timeSeriesMatrix}{description TBD}

\item{listOfParameters}{description TBD}
}
\value{
evaluated skew student t distribution fits
}
\description{
eval skew t distributions
}
\examples{
none

}

