% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-evaluateStudentTDistributions.R
\name{evaluateStudentTDistributions}
\alias{evaluateStudentTDistributions}
\title{eval t distributions}
\usage{
evaluateStudentTDistributions(timeSeriesMatrix, listOfParameters)
}
\arguments{
\item{timeSeriesMatrix}{description TBD}

\item{listOfParameters}{description TBD}
}
\value{
evaluated student t distribution fits
}
\description{
eval t distributions
}
\examples{
none

}

