% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-fitCopula.R
\name{fitCopula}
\alias{fitCopula}
\title{Fit a Vine Copula}
\usage{
fitCopula(dataObject)
}
\arguments{
\item{dataObject}{}
}
\value{
vineCopulaObject description TBD
}
\description{
Fit a Vine Copula
}
\examples{
none
}

