% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-fitMarginalResidual.R
\name{fitMarginalResidual}
\alias{fitMarginalResidual}
\title{Fit marginal distributions for noisy principal components - assumes student t distributions or normal distributions}
\usage{
fitMarginalResidual(timeSeriesMatrix, residualEigenValues, residualEigenVectors)
}
\arguments{
\item{timeSeriesMatrix}{xts time series matrix}

\item{residualEigenVectors}{noisy eigenvectors determined by dominantEigenvectors}

\item{resudualEigenValues}{noisy eigenvalues determined by dominantEigenvectors}
}
\value{
a list with normal or student t distribution fits for each component
}
\description{
Fit marginal distributions for noisy principal components - assumes student t distributions or normal distributions
}
\examples{
none

}

