% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-fitMarginalSignal.R
\name{fitMarginalSignal}
\alias{fitMarginalSignal}
\title{Fit signal distributions for signal principal components - assumes skew student t distributions}
\usage{
fitMarginalSignal(timeSeriesMatrix, signalEigenValues, signalEigenVectors)
}
\arguments{
\item{timeSeriesMatrix}{xts time series matrix}

\item{signalEigenValues}{signal eigenvalues determined by dominantEigenvectors}

\item{signalEigenVectors}{signal eigenvectors determined by dominantEigenvectors}
}
\value{
a list with skew student t distribution fits for each component
}
\description{
Fit signal distributions for signal principal components - assumes skew student t distributions
}
\examples{
none

}

