% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-fitNormalDistributions.R
\name{fitNormalDistributions}
\alias{fitNormalDistributions}
\title{fit normal distributions}
\usage{
fitNormalDistributions(timeSeriesMatrix)
}
\arguments{
\item{timeSeriesMatrix}{description TBD}
}
\value{
return normal distribution fits
}
\description{
fit normal distributions
}
\examples{
none

}

