% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-fitSkewStudentTDistributions.R
\name{fitSkewStudentTDistributions}
\alias{fitSkewStudentTDistributions}
\title{fit skew t distributions}
\usage{
fitSkewStudentTDistributions(timeSeriesMatrix)
}
\arguments{
\item{timeSeriesMatrix}{description TBD}
}
\value{
return skew t distribution fits
}
\description{
fit skew t distributions
}
\examples{
none

}

