% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-fitStudentTDistributions.R
\name{fitStudentTDistributions}
\alias{fitStudentTDistributions}
\title{fit t distributions}
\usage{
fitStudentTDistributions(timeSeriesMatrix)
}
\arguments{
\item{timeSeriesMatrix}{description TBD}
}
\value{
return t distribution fits
}
\description{
fit t distributions
}
\examples{
none

}

