% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Pricing-E-horizonPnLs.R
\name{horizonPnLs}
\alias{horizonPnLs}
\title{Calculate horizon price differences for use in portfolio optimization}
\usage{
horizonPnLs(simulatedHorizonPrices, horizonPriceVector)
}
\arguments{
\item{simulatedHorizonPrices}{is a matrix with simulated horizon prices, output from simulateDataToHorizon function}

\item{horizonPriceVector}{is a vector of prices used to calculate the differences}
}
\value{
a matrix of price differences
}
\description{
Calculate horizon price differences for use in portfolio optimization
}
\examples{
none

}

