% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-Pricing-E-horizonPrices.R
\name{horizonPrices}
\alias{horizonPrices}
\title{Calculate projected prices at investment horizon}
\usage{
horizonPrices(simulatedHorizonInvariants, horizonPriceVector)
}
\arguments{
\item{simulatedHorizonInvariants}{is a matrix of invariants, used to project prices, given a pricing function}

\item{horizonPriceVector}{is a vector of prices used to begin the projection from}
}
\value{
a matrix prices at the horizon
}
\description{
Calculate projected prices at investment horizon
}
\examples{
none

}

