% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-DataImport-E-importEquityPricesToXtsMatrix.R
\name{importEquityPricesToXtsMatrix}
\alias{importEquityPricesToXtsMatrix}
\title{Import symbols from Yahoo! Finance using getSymbols directly into a price matrix as xts object}
\usage{
importEquityPricesToXtsMatrix(symbolVector, startDate = "2000-01-01",
  endDate = Sys.Date(), priceType = "Adjusted", freq = "daily",
  keep = TRUE)
}
\arguments{
\item{symbolVector}{is a vector of symbols to download}

\item{startDate}{staring date of the matrix}

\item{endDate}{ending date of the matrix}

\item{priceType}{one of Open/High/Low/Close/Adjusted}

\item{freq}{one of daily/weekly/monthly/quarterly or yearly}

\item{keep}{logical, whether to keep the original OHLC xts matrices or to drop them}
}
\value{
an xts object with number of columns equal to length of symbolVector
}
\description{
Import symbols from Yahoo! Finance using getSymbols directly into a price matrix as xts object
}
\examples{
none

}

