% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07-Optimization-E-optimizeMeanVariancePortfolio.R
\name{optimizeMeanVariancePortfolio}
\alias{optimizeMeanVariancePortfolio}
\title{Optimize mean variance portfolio using Meucci's methods (holdings and simulated prices at the horizon) - brute force approach}
\usage{
optimizeMeanVariancePortfolio(meanVector, covarianceMatrix, horizonPriceVector,
  horizonPnLMatrix, wealthBalance, maxWeightConstraint, colNames, longOnly)
}
\arguments{
\item{meanVector}{is a vector of mean price changes}

\item{covarianceMatrix}{is a square, semi-positive definite covariance matrix of price changes}

\item{horizonPriceVector}{is a vector of security prices at the horizon}

\item{horizonPnLMatrix}{is a matrix of prices changes at the horizon}

\item{wealthBalance}{is the balance in the account (used to calculate holdings)}

\item{maxWeightConstraint}{is a list of symbols what will be subject to a max inverse volatility weight}

\item{colNames}{names of symbols}

\item{longOnly}{logical, whether to keep the original OHLC xts matrices or to drop them}
}
\value{
a list of quadratic programming efficient frontier results
}
\description{
Optimize mean variance portfolio using Meucci's methods (holdings and simulated prices at the horizon) - brute force approach
}
\examples{
none

}

