% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-plotDistributionFitStudentT.R
\name{plotDistributionFitStudentT}
\alias{plotDistributionFitStudentT}
\title{plot t distributions}
\usage{
plotDistributionFitStudentT(timeSeriesVector, listOfParameters)
}
\arguments{
\item{timeSeriesVector}{description TBD}

\item{listOfParameters}{description TBD}
}
\value{
plot
}
\description{
plot t distributions
}
\examples{
none

}

