% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88-Plot-E-plotOptimalStatistic.R
\name{plotOptimalStatistic}
\alias{plotOptimalStatistic}
\title{Plot time series object in ggplot}
\usage{
plotOptimalStatistic(optimizationResultsObject, statisticType,
  datePeriodType = "train", plotTitle = NULL, yTitle = NULL,
  xTitle = "Date", legendTitle = NULL, annualize = FALSE,
  interactive = FALSE)
}
\arguments{
\item{optimizationResultsObject}{a list containing rolling statistics for an optimization}

\item{statisticType}{character string, one of "opHoldings", "opWeights", "opER", "opSD", "opSharpe", "opIndex" for Optimal result; one of "mvHoldings", "mvWeights", "mvER", "mvSD", "mvSharpe", "mvIndex" for Minimum Variance result; the following are not available for this funciton: "ER" for expected return, "SD" for standard deviation, and "EF" for efficient frontier}

\item{datePeriodType}{character string, one of "train", "test", currently assumes one month ahead, defaults to "train"}

\item{plotTitle}{character string of plot title}

\item{yTitle}{character string of y-axis title}

\item{xTitle}{title for x-axis, defaults to "Date"}

\item{annualize}{logical that will annualize Expected Return, Standard Deviation or Sharpe ratio, if it's not applicable, it will be ignored; defaults to FALSE}

\item{interactive}{logical, whether or not to render in plotly for interactive use of plot; defaults to FALSE}
}
\value{
a timeseries plot (ggplot)
}
\description{
Plot time series object in ggplot
}
\examples{
none

}

