% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88-Plot-E-plotRiskReturnScatter.R
\name{plotRiskReturnScatter}
\alias{plotRiskReturnScatter}
\title{Plot time series object in ggplot}
\usage{
plotRiskReturnScatter(xtsTimeSeriesObject, plotTitle,
  yTitle = "Annualized Return", xTitle = "Annualized Risk",
  riskType = "SD", interactive = FALSE)
}
\arguments{
\item{xtsTimeSeriesObject}{an xts matrix of *returns*}

\item{plotTitle}{character string of plot title}

\item{yTitle}{character string of y-axis title; defaults to "Annualized Return"}

\item{xTitle}{title for x-axis, defaults to "Annualized Risk"}

\item{interactive}{logical, whether or not to render in plotly for interactive use of plot; defaults to FALSE}
}
\value{
a timeseries plot (ggplot)
}
\description{
Plot time series object in ggplot
}
\examples{
none

}

