% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88-Plot-E-plotTimeSeriesObject.R
\name{plotTimeSeriesObject}
\alias{plotTimeSeriesObject}
\title{Plot time series object in ggplot}
\usage{
plotTimeSeriesObject(xtsTimeSeriesObject, plotTitle = "Plot of Time Series",
  yTitle = "Time Series Value", yLabelType = "comma", xTitle = "Date",
  legendTitle = "Ticker", levels = "raw", interactive = FALSE)
}
\arguments{
\item{xtsTimeSeriesObject}{an xts matrix of *levels*}

\item{plotTitle}{character string of plot title}

\item{yTitle}{character string of y-axis title}

\item{yLabelType}{is a string, one of "dollar", "percent", "scientific", "comma"; defaults to comma}

\item{xTitle}{title for x-axis, defaults to "Date"}

\item{legendTitle}{title for legend, defaults to "Ticker"}

\item{levels}{character string, one of "linear", "continuous" or "raw"; if "raw", passes the timeSeries as-is for plotting, othwerise, will convert vector into an index (e.g. Prices) based on the selection of "linear" or "continuous" returns; defaults to "raw"}

\item{interactive}{logical, whether or not to render in plotly for interactive use of plot; defaults to FALSE}
}
\value{
a timeseries plot (ggplot)
}
\description{
Plot time series object in ggplot
}
\examples{
none

}

