% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retMatrixC.R
\name{retMatrixC}
\alias{retMatrixC}
\title{Calculate continuous (log) returns for a matrix}
\usage{
retMatrixC(priceMatrix)
}
\arguments{
\item{priceMatrix}{is an xts price matrix}
}
\value{
a continuous return xts matrix
}
\description{
Calculate continuous (log) returns for a matrix
}
\examples{
none

}

