% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retMatrixCtoL.R
\name{retMatrixCtoL}
\alias{retMatrixCtoL}
\title{Convert continuous (log) returns to linear returns for a matrix}
\usage{
retMatrixCtoL(CRetMatrix)
}
\arguments{
\item{CRetMatrix}{is a log return xts matrix}
}
\value{
a linear return xts matrix
}
\description{
Convert continuous (log) returns to linear returns for a matrix
}
\examples{
none

}

