% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retMatrixL.R
\name{retMatrixL}
\alias{retMatrixL}
\title{Calculate linear returns for a matrix}
\usage{
retMatrixL(priceMatrix)
}
\arguments{
\item{priceMatrix}{is an xts price matrix}
}
\value{
a linear return xts matrix
}
\description{
Calculate linear returns for a matrix
}
\examples{
none

}

