% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retMatrixLtoC.R
\name{retMatrixLtoC}
\alias{retMatrixLtoC}
\title{Convert linear to continuous (log) returns for a matrix}
\usage{
retMatrixLtoC(LRetMatrix)
}
\arguments{
\item{LRetMatrix}{is a log return xts matrix}
}
\value{
a linear return xts matrix
}
\description{
Convert linear to continuous (log) returns for a matrix
}
\examples{
none

}

