% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retVectorC.R
\name{retVectorC}
\alias{retVectorC}
\title{Calculate continuous (log) returns for a vector}
\usage{
retVectorC(timeseriesPriceVector)
}
\arguments{
\item{timeseriesPriceVector}{is an xts price vector}
}
\value{
a continuous return xts object
}
\description{
Calculate continuous (log) returns for a vector
}
\examples{
none

}

