% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retVectorCtoL.R
\name{retVectorCtoL}
\alias{retVectorCtoL}
\title{Convert log to linear returns for a vector}
\usage{
retVectorCtoL(timeseriesCRetVector)
}
\arguments{
\item{timeseriesCRetVector}{is an xts log return vector}
}
\value{
a linear return xts object
}
\description{
Convert log to linear returns for a vector
}
\examples{
none

}

