% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retVectorL.R
\name{retVectorL}
\alias{retVectorL}
\title{Calculate linear returns for a vector}
\usage{
retVectorL(timeseriesPriceVector)
}
\arguments{
\item{timeseriesPriceVector}{is an xts price vector}
}
\value{
a linear return xts object
}
\description{
Calculate linear returns for a vector
}
\examples{
none

}

