% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-retVectorLtoC.R
\name{retVectorLtoC}
\alias{retVectorLtoC}
\title{Convert linear to log returns for a vector}
\usage{
retVectorLtoC(timeseriesLRetVector)
}
\arguments{
\item{timeseriesLRetVector}{is an xts linear return vector}
}
\value{
a log-return xts object
}
\description{
Convert linear to log returns for a vector
}
\examples{
none

}

