% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-Projection-E-simulateDataToHorizon.R
\name{simulateDataToHorizon}
\alias{simulateDataToHorizon}
\title{Simulate prices to horizon based on fitted marginal distributions and copula - using Meucci's methods}
\usage{
simulateDataToHorizon(numSimulations, numPeriodsForward, vineCopulaFit,
  signalMarginalFit, residualMarginalFit,
  eigenVectors = diag((length(signalMarginalFit) +
  length(residualMarginalFit))))
}
\arguments{
\item{numSimulations}{is a scalar dictating number of simulations (rows) to conduct}

\item{numPeriodsForward}{is a scalar dictating number of periods forward (columns) the investment horizon is}

\item{vineCopulaFit}{is a vine copula object}

\item{signalMarginalFit}{is a list of marginal distribution fits for the signal risk factors}

\item{residualMarginalFit}{is a list of marginal distribution fits for the noisy risk factors}

\item{eigenVectors}{is a matrix of eigenVectors used to convert the horizon risk factors back into invariants}
}
\value{
a matrix of invariants projected to the investment horizon - to then be used in the pricing function
}
\description{
Simulate prices to horizon based on fitted marginal distributions and copula - using Meucci's methods
}
\examples{
none

}

