% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-toColList.R
\name{toColList}
\alias{toColList}
\title{convert matrix object, to a list of columns - used to enable use of lapply over apply(,1,FUN), which is significantly faster}
\usage{
toColList(matrixToConvert)
}
\arguments{
\item{matrixToConvert}{a numerical matrix}
}
\value{
a list of the columns of the matrix
}
\description{
convert matrix object, to a list of columns - used to enable use of lapply over apply(,1,FUN), which is significantly faster
}
\examples{
none

}

