% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-toDF.R
\name{toDF}
\alias{toDF}
\title{convert an xts object, where first column is a date vector, to a data frame object}
\usage{
toDF(rawXtsObject)
}
\arguments{
\item{rawXtsObject}{}
}
\value{
data.frame object
}
\description{
convert an xts object, where first column is a date vector, to a data frame object
}
\examples{
none

}

