% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-toRowList.R
\name{toRowList}
\alias{toRowList}
\title{convert matrix object, to a list of rows - used to enable use of lapply over apply(,1,FUN), which is significantly faster}
\usage{
toRowList(matrixToConvert)
}
\arguments{
\item{matrixToConvert}{a numerical matrix}
}
\value{
a list of the rows of the matrix
}
\description{
convert matrix object, to a list of rows - used to enable use of lapply over apply(,1,FUN), which is significantly faster
}
\examples{
none

}

