% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-General-E-toXTS.R
\name{toXTS}
\alias{toXTS}
\title{convert data frame object, where first column is a date vector, to an xts object}
\usage{
toXTS(rawDataFrame)
}
\arguments{
\item{rawDataFrame}{}
}
\value{
xts object
}
\description{
convert data frame object, where first column is a date vector, to an xts object
}
\examples{
none

}

