% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-DimensionReduction-E-tsEigenProduct.R
\name{tsEigenProduct}
\alias{tsEigenProduct}
\title{Convenience funciton to matrix multiply an xts object with  a set of eigenvectors while preserving the date-index}
\usage{
tsEigenProduct(timeSeriesMatrix, EigenVectors)
}
\arguments{
\item{timeSeriesMatrix}{is an xts time series matrix}

\item{EigenVectors}{is a matrix of eigenvectors}
}
\value{
an xts object
}
\description{
Convenience funciton to matrix multiply an xts object with  a set of eigenvectors while preserving the date-index
}
\examples{
none

}

