% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99-Date-E-createRollingTimeIndices.R
\name{createRollingTimeIndices}
\alias{createRollingTimeIndices}
\title{create index of time-stamps for vectorized functions}
\usage{
createRollingTimeIndices(initialTimeVector, endTimeDate = Sys.Date(),
  isExpanding = FALSE)
}
\arguments{
\item{initialTimeVector}{vector of dates YYYY-MM-DD}

\item{endTimeDate}{default to Sys.Date()}

\item{isExpanding}{logical, default FALSE}
}
\value{
vector of YYYY-MM dates
}
\description{
create index of time-stamps for vectorized functions
}
\examples{
none

}

