% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88-Plot-E-plotRiskReturnScatter.R
\name{plotRiskReturnScatter}
\alias{plotRiskReturnScatter}
\title{Plot time series object in ggplot}
\usage{
plotRiskReturnScatter(xtsTimeSeriesObject, plotTitle,
  yTitle = "Annualized Return", xTitle = "Annualized Risk",
  scale = "daily")
}
\arguments{
\item{xtsTimeSeriesObject}{an xts matrix of *returns*}

\item{plotTitle}{character string of plot title}

\item{yTitle}{character string of y-axis title; defaults to "Annualized Return"}

\item{xTitle}{title for x-axis, defaults to "Annualized Risk"}

\item{scale}{is a character string, one of "daily", "weekly", "monthly", "quarterly", "yearly"}
}
\value{
a timeseries plot (ggplot)
}
\description{
Plot time series object in ggplot
}
\examples{
none

}

