% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/88-Plot-E-plotTimeSeriesObject.R
\name{plotTimeSeriesObject}
\alias{plotTimeSeriesObject}
\title{Plot time series object in ggplot}
\usage{
plotTimeSeriesObject(xtsTimeSeriesObject, plotTitle, yTitle,
  yLabelType = scales::dollar, xTitle = "Date", levels = FALSE)
}
\arguments{
\item{xtsTimeSeriesObject}{an xts matrix of *levels*}

\item{plotTitle}{character string of plot title}

\item{yTitle}{character string of y-axis title}

\item{yLabelType}{is an object from the *scales* package, one of dollar, percent, scientific, comma; defaults to dollar}

\item{xTitle}{title for x-axis, defaults to "Date"}

\item{levels}{logical, if FALSE (default) indicates that the object contents are returns, if TRUE, indicates that the object contents are levels (e.g. Prices)}
}
\value{
a timeseries plot (ggplot)
}
\description{
Plot time series object in ggplot
}
\examples{
none

}

